Move sign_pkg_behavior into its own file
