+++ iscript/src/iscript/data/i_task_schema.json
@@ -92,7 +92,7 @@
                 "mac_notarize_part_1",
                 "mac_notarize_part_3",
                 "mac_single_file",
-                "mac_notarize_vpn",
+                "mac_sign_pkg",
                 "mac_sign_and_pkg_vpn",
                 "mac_notarize_single_file",
                 "mac_notarize_geckodriver"
